fix(ttl): TTL can be up to (2^31)-1

+++ src/session.rs
@@ -46,7 +46,7 @@ impl RedisSession {
     }
 
     /// Set time to live in seconds for session value
-    pub fn ttl(mut self, ttl: u16) -> Self {
+    pub fn ttl(mut self, ttl: i32) -> Self {
         Rc::get_mut(&mut self.0).unwrap().ttl = format!("{}", ttl);
         self
     }
